requirements.txt
uWSGI>=2.0.28
celery>=5.4.0
crispy-bootstrap5>=2024.2
Django>=5.1.5
django-cacheops>=7.1
django-celery-beat>=2.7.0
django-crispy-forms>=2.1
django-debug-toolbar>=5.0.1
django-environ>=0.12.0
django-extensions>=3.2.3
django-htmx>=1.22.0
django-redis>=5.4.0
ipython>=8.2.0
pillow>=10.3.0
psycopg2>=2.9.9
psycopg>=3.2.4
psycopg-pool>=3.2.4
redis>=5.2.1
django-celery-results>=2.5.1
dj-database-url>=2.3.0
django-allauth>=65.4.1
requests>=2.32.3
jwt>=1.3.1
django-compressor>=4.5.1
django-htmlmin>=0.11.0
pytz>=2024.2
django-unfold>=0.54.0